feat: Added functionality to Edit and Delete links
When you click the delet tag, the user can delte that selected restaurnt
from the database and is redirected to /restaurant page.
When the user clicks on the edit page and renames the restaurant.
They are redirected to '/restaurant' page.

Problem: once the user has submitted the new name. It does not redirect.
Has the following error: localhost didn’t send any data. ERR_EMPTY_RESPONSE

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch master
# Changes to be committed:
#	modified:   webserver.py
#
# Changes not staged for commit:
#	modified:   restaurantmenu.db
#
# Untracked files:
#	../../Vagrantfile.orig_4_6
#	../.vscode/
#